+++ core/amber/src/main/resources/python_udf/operators/texera_udf_operator_base.py
@@ -1,11 +1,9 @@
 import functools
 import logging
-import pickle
 from abc import ABC
 from typing import Dict, Optional, Tuple, List
 
 import pandas
-from sklearn.model_selection import train_test_split
 
 
 def exception(logger):
@@ -94,57 +92,6 @@ class TexeraUDFOperator(ABC):
         Executes when the input is exhausted, useful for some blocking execution like training.
         """
         pass
-
-
-class TexeraBlockingSupervisedTrainerOperator(TexeraUDFOperator):
-    logger = logging.getLogger("PythonUDF.TexeraBlockingSupervisedTrainerOperator")
-
-    @exception(logger)
-    def __init__(self):
-        super().__init__()
-        self._x = []
-        self._y = []
-        self._test_ratio = None
-        self._train_args = dict()
-        self._model_file_path = None
-
-    @exception(logger)
-    def input_exhausted(self, *args, **kwargs):
-        x_train, x_test, y_train, y_test = train_test_split(self._x, self._y, test_size=self._test_ratio, random_state=1)
-        model = self.train(x_train, y_train, **self._train_args)
-        with open(self._model_file_path, "wb") as file:
-            pickle.dump(model, file)
-
-        if x_test:
-            y_pred = self.test(model, x_test, y_test)
-            self.report(y_test, y_pred)
-
-    @exception(logger)
-    def accept(self, row: pandas.Series, nth_child: int = 0) -> None:
-        self._x.append(row[0])
-        self._y.append(row[1])
-
-    @staticmethod
-    @exception(logger)
-    def train(x_train, y_train, *args, **kwargs):
-        raise NotImplementedError
-
-    @staticmethod
-    @exception(logger)
-    def test(model, x_test, y_test, *args, **kwargs):
-        pass
-
-    @exception(logger)
-    def report(self, y_test, y_pred, *args, **kwargs):
-        from sklearn.metrics import classification_report
-        matrix = pandas.DataFrame(classification_report(y_test, y_pred, output_dict=True)).transpose()
-        matrix['class'] = [label for label, row in matrix.iterrows()]
-        cols = matrix.columns.to_list()
-        cols = [cols[-1]] + cols[:-1]
-        matrix = matrix[cols].round(3)
-        for index, row in matrix.iterrows():
-            if index != 1:
-                self._result_tuples.append(row)
 
 
 class TexeraBlockingUnsupervisedTrainerOperator(TexeraUDFOperator):
